+++ src/abstractdataframe/subset.jl
@@ -65,7 +65,10 @@ function assert_bool_vec(@nospecialize(fun))
     return function(x...)
         val = fun(x...)
         if !(val isa AbstractVector)
-            throw(ArgumentError("functions passed to `subset` must return an AbstractVector."))
+            throw(ArgumentError("functions passed to `subset`/`subset!` must return " *
+                                "an `AbstractVector` when subsetting a data frame " *
+                                "to ensure common mistakes in code are caught. Please " *
+                                "report an issue if you find this restriction inconvenient."))
         end
         return val
     end
@@ -88,11 +91,19 @@ function _get_subset_conditions(df::Union{AbstractDataFrame, GroupedDataFrame},
     conditions = Any[if a isa ColumnIndex
                          a => Symbol(:x, i)
                      elseif a isa Pair{<:Any, <:Base.Callable}
-                         first(a) => assert_bool_vec(last(a)) => Symbol(:x, i)
+                         # we require that vector is returned by the condition only
+                         # for AbstractDataFrame
+                         if df isa GroupedDataFrame
+                             first(a) => last(a) => Symbol(:x, i)
+                         else
+                             @assert df isa AbstractDataFrame
+                             first(a) => assert_bool_vec(last(a)) => Symbol(:x, i)
+                         end
                      else
                          throw(ArgumentError("condition specifier $a is not supported by `subset`"))
                      end for (i, a) in enumerate(cs_vec)]
-    isempty(conditions) && throw(ArgumentError("at least one condition must be passed"))
+    # we handle empty conditions case outside of this function
+    @assert !isempty(conditions)
 
     if df isa AbstractDataFrame
         df_conditions = select(df, conditions..., copycols=!(df isa DataFrame))
@@ -125,7 +136,17 @@ function _get_subset_conditions(df::Union{AbstractDataFrame, GroupedDataFrame},
         end
     end
 
-    @assert eltype(cond) === Bool
+    # we special case 0-length cond, as in this case broadcasting does not
+    # guarantee setting a proper eltype for the result
+    if isempty(cond)
+        if eltype(cond) !== Bool
+            throw(ArgumentError("passed conditions produce $(eltype(cond)) " *
+                                "as element type of the result while only " *
+                                "Bool is allowed."))
+        end
+    else
+        @assert eltype(cond) === Bool
+    end
     return cond
 end
 
@@ -136,21 +157,23 @@ end
 
 Return a copy of data frame `df` or parent of `gdf` containing only rows for
 which all values produced by transformation(s) `args` for a given row are
-`true`. All transformations must produce vectors containing `true` or `false`
-(and optionally `missing` if `skipmissing=true`).
+`true`. All transformations must produce vectors containing `true` or `false`.
+When the first argument is a `GroupedDataFrame`, transformations are also
+allowed to return a single `true` or `false` value, which results in including
+or excluding a whole group.
+
+If `skipmissing=false` (the default) `args` are required to produce results
+containing only `Bool` values. If `skipmissing=true`, additionally `missing` is
+allowed and it is treated as `false` (i.e. rows for which one of the conditions
+returns `missing` are skipped).
 
 Each argument passed in `args` can be any specifier following the rules
 described for [`select`](@ref) with the restriction that:
 * specifying target column name is not allowed as `subset` does not create new
   columns;
 * every passed transformation must return a scalar or a vector (returning
-  `AbstractDataFrame`, `NamedTuple`, `DataFrameRow` or `AbstractMatrix`
-  is not supported).
-
-If `skipmissing=false` (the default) `args` are required to produce vectors
-containing only `Bool` values. If `skipmissing=true`, additionally `missing` is
-allowed and it is treated as `false` (i.e. rows for which one of the conditions
-returns `missing` are skipped).
+  `AbstractDataFrame`, `NamedTuple`, `DataFrameRow` or `AbstractMatrix` is not
+  supported).
 
 If `view=true` a `SubDataFrame` view  is returned instead of a `DataFrame`.
 
@@ -226,6 +249,14 @@ julia> subset(groupby(df, :y), :v => x -> x .> minimum(x))
 ─────┼─────────────────────────────────────
    1 │     2  false   true     true      2
    2 │     4  false  false  missing     12
+
+julia> subset(groupby(df, :y), :v => x -> minimum(x) > 5)
+2×5 DataFrame
+ Row │ id     x      y      z        v
+     │ Int64  Bool   Bool   Bool?    Int64
+─────┼─────────────────────────────────────
+   1 │     3   true  false  missing     11
+   2 │     4  false  false  missing     12
 ```
 """
 function subset(df::AbstractDataFrame, @nospecialize(args...);
@@ -264,8 +295,15 @@ end
 
 Update data frame `df` or the parent of `gdf` in place to contain only rows for
 which all values produced by transformation(s) `args` for a given row is `true`.
-All transformations must produce vectors containing `true` or `false` (and
-optionally `missing` if `skipmissing=true`).
+All transformations must produce vectors containing `true` or `false`. When the
+first argument is a `GroupedDataFrame`, transformations are also allowed to
+return a single `true` or `false` value, which results in including or excluding
+a whole group.
+
+If `skipmissing=false` (the default) `args` are required to produce results
+containing only `Bool` values. If `skipmissing=true`, additionally `missing` is
+allowed and it is treated as `false` (i.e. rows for which one of the conditions
+returns `missing` are skipped).
 
 Each argument passed in `args` can be any specifier following the rules
 described for [`select`](@ref) with the restriction that:
@@ -275,13 +313,8 @@ described for [`select`](@ref) with the restriction that:
   `AbstractDataFrame`, `NamedTuple`, `DataFrameRow` or `AbstractMatrix` is not
   supported).
 
-If `skipmissing=false` (the default) `args` are required to produce vectors
-containing only `Bool` values. If `skipmissing=true`, additionally `missing` is
-allowed and it is treated as `false` (i.e. rows for which one of the conditions
-returns `missing` are skipped).
-
-If `ungroup=false` the passed `GroupedDataFrame` `gdf` is updated (preserving the order
-of its groups) and returned.
+If `ungroup=false` the passed `GroupedDataFrame` `gdf` is updated (preserving
+the order of its groups) and returned.
 
 If `GroupedDataFrame` is subsetted then it must include all groups present in
 the `parent` data frame, like in [`select!`](@ref). In this case the passed
@@ -367,6 +400,19 @@ julia> df
      │ Int64  Bool   Bool   Bool?    Int64
 ─────┼─────────────────────────────────────
    1 │     2  false   true     true      2
+   2 │     4  false  false  missing     12
+
+julia> df = DataFrame(id=1:4, x=[true, false, true, false], y=[true, true, false, false],
+                      z=[true, true, missing, missing], v=[1, 2, 11, 12]);
+
+julia> subset!(groupby(df, :y), :v => x -> minimum(x) > 5);
+
+julia> df
+2×5 DataFrame
+ Row │ id     x      y      z        v
+     │ Int64  Bool   Bool   Bool?    Int64
+─────┼─────────────────────────────────────
+   1 │     3   true  false  missing     11
    2 │     4  false  false  missing     12
 ```
 """